+++ src/main/java/domain/Race.java
@@ -1,4 +1,4 @@
-package service;
+package domain;
 
 public interface Race {
