+++ src/main/java/domain/vo/Car.java
@@ -1,4 +1,4 @@
-package service.domain;
+package domain.vo;
 
 public class Car {
