+++ src/main/java/domain/vo/RaceCondition.java
@@ -1,4 +1,4 @@
-package service.domain;
+package domain.vo;
 
 import java.util.ArrayList;
