+++ src/test/java/service/domain/CarTest.java
@@ -1,5 +1,6 @@
 package service.domain;
 
+import domain.vo.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
