+++ src/test/java/view/MonitorViewTest.java
@@ -3,7 +3,7 @@ package view;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import service.domain.Car;
+import domain.vo.Car;
 
 import static org.junit.jupiter.api.Assertions.*;
